+++ basis/newtoncotes/main_test.go
@@ -12,7 +12,7 @@ func assertEqual(expected, actual interface{}, t *testing.T) {
 }
 
 func TestComputeOrders(t *testing.T) {
-	expectedOrders := [][]uint16{{0}, {0, 2}, {1, 3}, {1, 3, 5, 7}}
+	expectedOrders := [][]uint32{{0}, {0, 2}, {1, 3}, {1, 3, 5, 7}}
 
 	basis := New()
 
@@ -25,7 +25,7 @@ func TestComputeOrders(t *testing.T) {
 
 func TestComputeNodes(t *testing.T) {
 	levels := []uint8{0, 1, 1, 2, 2, 3, 3, 3, 3}
-	orders := []uint16{0, 0, 2, 1, 3, 1, 3, 5, 7}
+	orders := []uint32{0, 0, 2, 1, 3, 1, 3, 5, 7}
 
 	expectedNodes := []float64{0.5, 0, 1, 0.25, 0.75, 0.125, 0.375, 0.625, 0.875}
 
@@ -37,10 +37,10 @@ func TestComputeNodes(t *testing.T) {
 
 func TestComputeChildren(t *testing.T) {
 	levels := []uint8{0, 1, 1, 2, 2, 3, 3, 3, 3}
-	orders := []uint16{0, 0, 2, 1, 3, 1, 3, 5, 7}
+	orders := []uint32{0, 0, 2, 1, 3, 1, 3, 5, 7}
 
 	expectedLevels := []uint8{1, 1, 2, 2, 3, 3, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4}
-	expectedOrders := []uint16{0, 2, 1, 3, 1, 3, 5, 7, 1, 3, 5, 7, 9, 11, 13, 15}
+	expectedOrders := []uint32{0, 2, 1, 3, 1, 3, 5, 7, 1, 3, 5, 7, 9, 11, 13, 15}
 
 	basis := New()
 	actualLevels, actualOrders := basis.ComputeChildren(levels, orders)
